+++ blogDataService/src/main/java/com/finalProject/blogDataService/model/Post.java
@@ -12,7 +12,7 @@ public class Post {
 
     private String section;
 
-    private Integer userId;
+    private Long userId;
 
     private String description;
 
@@ -22,7 +22,7 @@ public class Post {
     public Post() {
     }
 
-    public Post(String title, LocalDate date, String section, Integer userId, String description, String urlImagen) {
+    public Post(String title, LocalDate date, String section, Long userId, String description, String urlImagen) {
         setTitle(title);
         setDate(date);
         setSection(section);
@@ -79,11 +79,11 @@ public class Post {
         this.section = section;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 }
