+++ blogDataService/src/main/java/com/finalProject/blogDataService/service/interfaces/IBlogService.java
@@ -3,5 +3,5 @@ package com.finalProject.blogDataService.service.interfaces;
 import com.finalProject.blogDataService.model.Blog;
 
 public interface IBlogService {
-    Blog getBlog(Integer id);
+    Blog getBlog(Long id);
 }
