+++ postInfoService /src/main/java/com/finalProject/postInfoService/repository/PostRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PostRepository extends JpaRepository<Post,Integer> {
-    List<Post> findByuserId(Integer userId);
+public interface PostRepository extends JpaRepository<Post, Integer> {
+    List<Post> findByuserId(Long userId);
 }
